requirements.txt
streamlit==1.41.1
mediapipe==0.10.21
opencv-python-headless==4.11.0.86
numpy==2.2.3
Pillow==11.1.0
scikit-image==0.25.1
replicate==1.0.4
python-dotenv==1.0.1
